Don't checkin generated reflectable stubs.
`reflectable` generates a metadata file named
`libname.reflectable.dart` which contains all of the generated
reflection data. This isn't meant to be under source control and is
generated when the package builds.

This removes it and adds an ignore rule to filter out any future
reflectable metadata libraries that happen to be generated in the
future.