Manually embed types into interpreter.
The dart-lang/reflectable library which is what we were using to
generate what methods and fields and constructors were available to a
class based on a static analysise of the codebase had two significant
drawbacks.

1) It adds a build step to the development cycle which introduces
quite a bit of uncertainty and instability to the process.
2) It doesn't appear to work for some pretty fundamental flutter
classes.

This change bins the reflectable approach in favor of a manual way of
specifying which fields, methods, and constructors are embedded into
the interpreter.

While this is more verbose, since we're effectively maintaining by
hand what the generator was generating for us we're also

1. very explicit about what is a available and what's not, and as a
result have a very high degree of control over the outcome.
2. completely independent of any build step in both development and
publish.
3. able to move forward without being blocked on the reflectable apis.
4. No need to call some `initializeReflectable()` method in your
`main`.

Notable other changes:

- `DartFn` is now just a simple `typedef` not a class.
the property access macro `::` is removed since property access is
just another kind of method.
- array-like access has been added to the dart arguments object.