Make a handy closure syntax that is callable from both fs and dart
Defining functions in FlutterScript is handy, but currently it's only
really convenient to invoke them from within FlutterScript. Sometimes
what we want is to capture a function that is defined in
FlutterScript, but have it _also_ be callable from Dart code.

This change introduces a new syntax for defining closures, or
"lambdas" in FlutterScript, and if this syntax is used, then functions
of that type will also be callable from within Dart. This is achieved
by introducing the `FlutterScriptFn` class which is like a normal LISP
closure except that it captures a reference to the interpreter which
gave rise to it, so that it can be used to invoke it without having to
keep track of the interpreter.

The function syntax is based on fat arrow syntax from Dart and
JavaScript, bmut in all other ways looks like a lambda expression

```lisp
(=> (x y) (+ x y))
```

will define a function that takes two parameters, and adds them
together. Not only can you call this from FlutterScript, but you can
also call it from within Dart:

```dart
FlutterScriptFn add = await interpreter.eval('(=> (x y) (+ x y))');
```

and it will be callable from Dart by passing in a list of the
arguments:

```dart
add([5, 10]); //=> 15
add([2, 3]); //=> 5